Move parameters to ButtonStateStyle
