+++ component/compose-stylesheet-button/src/commonMain/kotlin/com.moriatsushi.compose.stylesheet.button/ButtonStateStyle.kt
@@ -3,12 +3,15 @@ package com.moriatsushi.compose.stylesheet.button
 import androidx.compose.runtime.Immutable
 import com.moriatsushi.compose.stylesheet.component.ComponentCommonStyle
 import com.moriatsushi.compose.stylesheet.content.ContentStyle
+import com.moriatsushi.compose.stylesheet.indication.IndicationStyle
 
 /**
  * A style for each [Button] state, such as pressed, focused, etc.
  */
 @Immutable
 sealed interface ButtonStateStyle {
+    val layout: ButtonLayout
+    val indication: IndicationStyle?
     val contentStyle: ContentStyle
     val commonStyle: ComponentCommonStyle
 
@@ -24,15 +27,21 @@ internal fun ButtonStateStyle(builder: ButtonStateStyleBuilder.() -> Unit): Butt
     ButtonStateStyleBuilder().apply(builder).buildStateStyle()
 
 internal fun ButtonStateStyle(
+    layout: ButtonLayout = ButtonLayout.Default,
+    indication: IndicationStyle? = null,
     commonStyle: ComponentCommonStyle = ComponentCommonStyle.Default,
     contentStyle: ContentStyle = ContentStyle.Default,
 ): ButtonStateStyle = ButtonStateStyleImpl(
+    layout = layout,
+    indication = indication,
     commonStyle = commonStyle,
     contentStyle = contentStyle,
 )
 
 @Immutable
 private data class ButtonStateStyleImpl(
-    override val commonStyle: ComponentCommonStyle = ComponentCommonStyle.Default,
-    override val contentStyle: ContentStyle = ContentStyle.Default,
+    override val layout: ButtonLayout,
+    override val indication: IndicationStyle?,
+    override val commonStyle: ComponentCommonStyle,
+    override val contentStyle: ContentStyle,
 ) : ButtonStateStyle
